% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see}
\alias{see}
\title{Just prints a vector in a nicer format}
\usage{
see(vec)
}
\arguments{
\item{vec}{Vec to see}
}
\description{
Prints a vector with entries separated by two newlines and a hashtagline.
Easier to read, i think.
}
\examples{
see(rownames(mtcars)) 
}
\keyword{convenience}
