% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsToDates.R
\name{varsToDates}
\alias{varsToDates}
\title{Merges Variables Into a Date Variable}
\usage{
varsToDates(varlist, fixNaDay = TRUE, fixNaMonth = FALSE)
}
\arguments{
\item{varlist}{list of date variables, like}
}
\description{
A function that merges several variables containing year, month and day into a single
date variable. The function can also substitute missing values with means, like middle
of month and so on.
}
\examples{
df$date <- varsToDates(list(year = df$year,month = df$month,day = df$day) 
}
\keyword{dates}
